// client/model/ruleOperationDto.ts
/**
 * Draas
 * Draas OAS Documentation
 *
 * The version of the OpenAPI document: 0.0.1
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface RuleOperationDto {
    existingRuleId: string;
    operation: string;
    destinationGroupId: string;
    ruleIndex: number;
    name: string;
}